requirements.txt
asgiref==3.3.1
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.4
django-allauth==0.44.0
django-appconf==1.0.4
django-cors-headers==3.6.0
django-custom-user==0.7
django-environ==0.4.5
django-filter==2.4.0
django-imagekit==4.0.2
django-import-export==2.4.0
django-ipware==3.0.2
django-jsonfield-schema==0.3.0
django-model-utils==4.1.1
django-rest-auth==0.9.5
django-storages==1.11
django-user-agents==0.4.0
django-webpack-loader==0.7.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
ecdsa==0.16.1
et-xmlfile==1.0.1
gunicorn==20.0.4
http-ece==1.1.0
idna==2.10
importlib-metadata==3.3.0
jdcal==1.4.1
Jinja2==2.11.2
jsonschema==3.2.0
MarkupPy==1.14
MarkupSafe==1.1.1
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.5
pilkit==2.0
Pillow==8.0.1
psycopg2==2.7
py-vapid==1.7.1
pyactiveresource==2.2.1
pycparser==2.20
PyJWT==1.7.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
pywebpush==1.11.0
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
sentry-sdk==0.19.5
ShopifyAPI==8.0.1
six==1.15.0
sqlparse==0.3.1
tablib==3.0.0
typing-extensions==3.7.4.3
ua-parser==0.10.0
urllib3==1.26.2
user-agents==2.2.0
wincertstore==0.2
xlrd==2.0.1
xlwt==1.3.0
zipp==3.4.0
